requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0

# For the MCP server (Python 3.10+ required)
mcp>=1.2.0 